Moved two functions in utils
